refactor: extract form filling logic
